move dynamicBody and muiTheme into _utils folder
-cleans up file structure
